requirements.txt
PyBoy
python-dotenv
discord
discord.py
